% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_entries.R
\name{all_entries}
\alias{all_entries}
\title{All Entry Scraper}
\usage{
all_entries(entries)
}
\arguments{
\item{entries}{a list passed from getEntries}
}
\value{
returns a dataframe with relevant information encoded in the entries such as author, rating, follows, and, date published.
}
\description{
This function Scrapes all of the pages in a particular fandom and turns it into a dataframe with information
}
\examples{
all_entries(getEntries(getPages(getUrl("NCIS", "tv"))))
}
