% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChapter.R
\name{getChapter}
\alias{getChapter}
\title{Scrape a Chapter of Desired Fanfiction}
\usage{
getChapter(returns)
}
\arguments{
\item{returns}{A list generated by getTitles including the fanfiction.net url and href tag of associated fanfic.}
}
\value{
The raw text scraped from a chapter of fanfiction
}
\description{
\code{getChapter.absentfan} Helper function of getChapter scraping a single chapter from specific entry in fanfiction.net.
}
\examples{
getChapter(list(url="https://www.fanfiction.net/s/13084808/1/Modern-Marauders",href="/s/13084808/1/Modern-Marauders"))
}
