% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntries.R
\name{getEntries}
\alias{getEntries}
\title{Scrapes a page of entries}
\usage{
getEntries(returns)
}
\arguments{
\item{returns}{a list returned by getPages}
}
\value{
returns a list of scraped entries from one page in a raw
   character string.
}
\description{
this function uses the information from getEntries
   in order to scrape a single page of data about stories
}
\examples{
getEntries(list(noPages=FALSE, page_use="/cartoon/Invader-Zim/?&srt=1&r=103&p=",
page_num=11, url="https://www.fanfiction.net/cartoon/Invader-Zim/", max.entries=NA))
}
