% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullStory.R
\name{getFullStory}
\alias{getFullStory}
\title{Scraper of Chapter Content in Fanfiction}
\usage{
getFullStory(story, type, title)
}
\arguments{
\item{story}{Name of fandom being searched. Ex, "Harry Potter."}

\item{type}{Whether the fandom is from the media type of anime, book, cartoon, comic, game, misc, movie, play, or tv.}

\item{title}{The title of the specific fanfic post to be scraped.}
}
\value{
A data frame with two columns; the raw text of the scraped entries, and the chapter designations for each text vector.
}
\description{
\code{getFullStory.absentfan} Scrapes fanfiction.net for a specific fanfiction entry (ie, user generated story)
based on the fandom (ie, general story narrative), type, and the verbatim title of the story.
}
\examples{
getFullStory("Harry Potter", "book", "Modern Marauders")
}
