% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPages.R
\name{getPages}
\alias{getPages}
\title{Get information on the quantity and structure of fandom pages}
\usage{
getPages(returns)
}
\arguments{
\item{returns}{list passed from getUrl function}
}
\value{
returns a list with 4 elements
   a boolean that returns TRUE if there are no pages associated with the URL
   a string that gives the pattern of the URLs of the pages
   a numeric with the total number of pages
   a string with the url passed to the function
   max.entries passed from getUrl
}
\description{
This function takes the output of the getURL function
   and generates information based on the amount of pages and the
   structure of the site.
}
\examples{
getPages(list(url="https://www.fanfiction.net/cartoon/Invader-Zim/", max.entries=NA))
getPages(list(url="https://www.fanfiction.net/game/Kingdom-Hearts/", max.entries=10))
}
