% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPagesTitle.R
\name{getPagesTitle}
\alias{getPagesTitle}
\title{Get Pages to Iterate When Scraping Specific Story}
\usage{
getPagesTitle(returns)
}
\arguments{
\item{returns}{A list produced by \code{getUrlTitle.absentfan} including the full url of the associated fandom and the title of the fanfic being retrieved.}
}
\value{
A list including noPages (bool whether there are multiple pages of entries to scrape), page_use (href, if needed), page_num (total number of pages to iterate), url, and story title.
}
\description{
\code{getPagesTitle.absentfan} Helper function of getChapter to get the total number of pages to iterate through
if the catalogue of entries is multiple pages.
}
\examples{
getPagesTitle(list(url="https://www.fanfiction.net/book/Harry-Potter/", title="Modern Marauders"))
}
