% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTitles.R
\name{getTitles}
\alias{getTitles}
\title{Get url of Specific Story to be Scraped}
\usage{
getTitles(returns)
}
\arguments{
\item{returns}{A list generated by \code{getPagesTitle.absentfan} including noPages (bool whether there are multiple pages of entries to scrape), page_use (href, if needed), page_num (total number of pages to iterate), url, and story title.}
}
\value{
A list including a url (of fanfiction.net) and the href of the story being queried.
}
\description{
\code{getTitles.absentfan} Helper function of getChapter iterating through all entry titles in fandom listing
in search of title queried by user.
}
\examples{
getTitles(list(noPages=FALSE, page_use="/book/Harry-Potter/?&srt=1&r=103&p=",page_num=11,url="https://www.fanfiction.net/book/Harry-Potter/", title="Modern Marauders"))
}
