% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl.R
\name{getUrl}
\alias{getUrl}
\title{Get URL of a Fandom}
\usage{
getUrl(story, type, max.entries = NA)
}
\arguments{
\item{story}{Which Fandom would you like the URL for?}

\item{type}{Which type of media? Must be one of the following
("anime", "book", "cartoon", "comic", "game", "misc", "tv")}

\item{max.entries}{How many Entries to scrape?}
}
\value{
returns a list with 2 entries
   the URL of the fandom page as a string
   max.entries
}
\description{
For a given fandom in a certain medium,
   this function gives the URL of its story page.
}
\examples{
getUrl("Zoolander","movie")
getUrl("Kingdom Hearts","game", max.entries=10)
}
