% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlTitle.R
\name{getUrlTitle}
\alias{getUrlTitle}
\title{Get URL of Specific Story}
\usage{
getUrlTitle(story, type, title)
}
\arguments{
\item{story}{Name of fandom being searched. Ex, "Harry Potter."}

\item{type}{Whether the fandom is from the media type of anime, book, cartoon, comic, game, misc, movie, play, or tv.}

\item{title}{The title of the specific fanfic post to be scraped.}
}
\value{
A list including the full url and the title of the fanfic being retrieved.
}
\description{
\code{getUrlTitle.absentfan} Helper function of getChapter retrieving the full url of fandom from fanfiction.net while passing down the title
of a story to be scraped by later functions.
}
\examples{
getUrlTitle("Harry Potter", "book", "Modern Marauders")
}
