% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeAllEntries.R
\name{scrapeAllEntries}
\alias{scrapeAllEntries}
\title{Fanfiction Data Scraper}
\usage{
scrapeAllEntries(story, type, max.entries = NA)
}
\arguments{
\item{story}{Which Fandom would you like to scrape?}

\item{type}{What media type is it?}

\item{max.entries}{How many Entries would you like to scrape?
If NA, scrape all of the entries.}
}
\value{
returns a dataframe with details about the entries scraped
}
\description{
Scrapes the data on stories on Fanfiction.net by fandom.
}
\examples{
scrapeAllEntries("Harry Potter", "book", max.entries=11)
}
