% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storyPrint.R
\name{storyPrint}
\alias{storyPrint}
\title{Chapter to HTML Output}
\usage{
storyPrint(story, type, title, browse = FALSE)
}
\arguments{
\item{story}{Name of fandom being searched. Ex, "Harry Potter."}

\item{type}{Whether the fandom is from the media type of anime, book, cartoon, comic, game, misc, movie, play, or tv.}

\item{title}{The title of the specific fanfic post to be scraped.}

\item{browse}{If \code{TRUE}, use external browser}
}
\value{
path to HTML file
}
\description{
Generate an easily readable html document with a fully scraped fanfiction retrieved from fanfiction.net.
}
\examples{
storyPrint("Harry Potter", "book", "Modern Marauders")
}
