% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeMedia.R
\docType{data}
\name{typeMedia}
\alias{typeMedia}
\title{Type Media Dataset}
\format{A list with 9 elements
\describe{
  \item{anime}{Anime Fandoms}
  \item{book}{Literary Fandoms}
  \item{cartoon}{Cartoon Fandoms}
  \item{comic}{Comic Fandoms}
  \item{game}{Video Game Fandoms}
  \item{misc}{Miscellanious Fandoms}
  \item{movie}{Cinematic Fandoms}
  \item{play}{Theatrical Fandoms}
  \item{tv}{Television Fandoms}
}}
\source{
Fanfiction.net
}
\usage{
typeMedia
}
\description{
This data was scraped from fanfiction.net on December 12, 2018.
}
\details{

}
\examples{
lapply(typeMedia, nrow)
}
\keyword{datasets}
