% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTypeMedia.R
\name{updateTypeMedia}
\alias{updateTypeMedia}
\title{Update Href tags for Type Media}
\usage{
updateTypeMedia()
}
\value{
list with all href tags for each type of media
}
\description{
This function scrapes each media page of fanfiction.net
   and returns a list of the href tags of each fandom. This takes a
   few minutes.
}
\details{
Use this function to update the href tags only if you believe
   you are missing a very recent story, as the tags do not change much
   if at all from day to day.
}
\examples{
typeMedia <- updateTypeMedia()
}
